auth-server: increase rate limit on settlement (#168)
